Cleanup build process for pre-codegen versions of react-native.
