--- android/src/fabric/java/com/swmansion/rnscreens/StateWrapper.kt
@@ -1,12 +1,0 @@
-package com.swmansion.rnscreens
-
-import com.facebook.react.bridge.ReadableMap
-import com.facebook.react.uimanager.FabricViewStateManager
-
-class StateWrapper {
-    companion object {
-        fun getStateData(state: FabricViewStateManager): ReadableMap? {
-            return state.getStateData()
-        }
-    }
-}
