+++ android/src/main/java/com/swmansion/rnscreens/Screen.kt
@@ -14,14 +14,11 @@ import com.facebook.react.bridge.ReactContext
 import com.facebook.react.bridge.ReadableMap
 import com.facebook.react.bridge.WritableMap
 import com.facebook.react.bridge.WritableNativeMap
-import com.facebook.react.uimanager.FabricViewStateManager
 import com.facebook.react.uimanager.PixelUtil
 import com.facebook.react.uimanager.UIManagerModule
-import kotlin.math.abs
 
 @SuppressLint("ViewConstructor")
-class Screen constructor(context: ReactContext?) : ViewGroup(context), FabricViewStateManager.HasFabricViewStateManager {
-    private val mFabricViewStateManager: FabricViewStateManager = FabricViewStateManager()
+class Screen constructor(context: ReactContext?) : FabricEnabledViewGroup(context) {
 
     var fragment: ScreenFragment? = null
     var container: ScreenContainer<*>? = null
@@ -89,40 +86,6 @@ class Screen constructor(context: ReactContext?) : ViewGroup(context), FabricVie
                         ?.updateNodeSize(id, width, height)
                 }
             })
-    }
-
-    private fun updateScreenSizeFabric(width: Int, height: Int) {
-        updateState(width, height)
-    }
-
-    @UiThread
-    fun updateState(width: Int, height: Int) {
-        val realWidth: Float = PixelUtil.toDIPFromPixel(width.toFloat())
-        val realHeight: Float = PixelUtil.toDIPFromPixel(height.toFloat())
-
-        // Check incoming state values. If they're already the correct value, return early to prevent
-        // infinite UpdateState/SetState loop.
-        val currentState: ReadableMap? = StateWrapper.getStateData(getFabricViewStateManager())
-        if (currentState != null) {
-            val delta = 0.9f
-            val stateFrameHeight: Float = if (currentState.hasKey("frameHeight")) currentState.getDouble("frameHeight").toFloat() else 0f
-            val stateFrameWidth: Float = if (currentState.hasKey("frameWidth")) currentState.getDouble("frameWidth").toFloat() else 0f
-            if (abs(stateFrameWidth - realWidth) < delta &&
-                abs(stateFrameHeight - realHeight) < delta
-            ) {
-                return
-            }
-        }
-        mFabricViewStateManager.setState {
-            val map: WritableMap = WritableNativeMap()
-            map.putDouble("frameWidth", realWidth.toDouble())
-            map.putDouble("frameHeight", realHeight.toDouble())
-            map
-        }
-    }
-
-    override fun getFabricViewStateManager(): FabricViewStateManager {
-        return mFabricViewStateManager
     }
 
     val headerConfig: ScreenStackHeaderConfig?
