+++ android/src/paper/java/com/facebook/react/viewmanagers/RNSScreenStackManagerInterface.java
